refactor: move runner markers and scoring pipeline from engine to analysis
