+++ tools/import_baseline.txt
@@ -22,7 +22,6 @@ src/quodeq/data/fs/report_parser/runs.py:216:services
 src/quodeq/data/projection/grade_projector.py:132:services
 src/quodeq/data/projection/grade_projector.py:20:services
 src/quodeq/data/sqlite/state_store.py:272:services
-src/quodeq/engine/scoring_pipeline.py:10:services
 src/quodeq/llm_bridge/_providers.py:6:analysis
 src/quodeq/services/_external_jobs.py:22:analysis
 src/quodeq/services/_violations_jsonl.py:12:analysis
